issue-634: refactor(coach): extract coach_screen chat/composer/credits widgets (#634)
Reduce coach_screen.dart 2008 -> 1418 lines by moving the already-separated
presentational widget classes into lib/widgets/coach/ (verbatim, no
behavioural change; private classes referenced by the screen promoted to
public, Dart privacy is per-file):

- coach_chat_bubble.dart   (ChatBubble + private _BubbleAvatar + MessageBody)
- coach_composer.dart      (ComposerInputField + SendButton)
- coach_typing_dots.dart   (CoachTypingDots)
- coach_credit_packs_sheet.dart (CreditPacksSheet)

Removed now-unused imports (flutter_markdown, app_theme). Residual >500 is the
_CoachScreenState body; clean reduction needs #632 (per #634 migration note).
Verified: analyze clean on changed files; flutter test test/screens/ green
(incl. coach_offline_and_fixes / coach_insight / coach_mode_recommender).

Part of #634 (coach; meal_planner + dashboard remain).
